+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.25-Release] 2020-12-18
+### Changed
+- Added back our more orange combo point color, the red was never actually intended.
+- The usual amount of back-end updates.
+
 ## [2.1.24-Release] 2020-12-16
 - Back-end updates.